Refactor + added case generation to appeals
