+++ src/plugins/appeals/unbanAppeal.ts
@@ -1,6 +1,7 @@
 import {useClient} from "../../hooks";
 import {useEvent} from "../../hooks";
 import {Interaction, Message} from "discord.js";
+import { unbanWithBlame } from "../moderation/unbanGenerator";
 
 useEvent("interactionCreate", async (interaction: Interaction) => {
     if (!interaction.isButton()) {
@@ -17,8 +18,7 @@ useEvent("interactionCreate", async (interaction: Interaction) => {
     await interaction.deferReply({ephemeral: true});
     const {client} = useClient();
     try {
-        const rApple = await client.guilds.fetch("332309672486895637");
-        await rApple.bans.remove(userId);
+        unbanWithBlame(interaction.user, userId, "Ban appealed successfully.")
         interaction.editReply(`Unbanned <@${userId}>`);
     } catch {
         interaction.reply("Could not unban user");
